+++ tap_zendesk_chat/sync.py
@@ -1,0 +1,33 @@
+from .utils import load_schema
+import singer
+from singer import metadata
+from . import streams as streams_
+
+
+
+
+def output_schema(stream):
+    schema = load_schema(stream.tap_stream_id)
+    singer.write_schema(stream.tap_stream_id, schema, stream.pk_fields)
+
+
+def is_selected(stream):
+    mdata = metadata.to_map(stream.metadata)
+    return metadata.get(mdata, (), 'selected')
+
+def sync(ctx):
+    currently_syncing = ctx.state.get("currently_syncing")
+    start_idx = streams_.all_stream_ids.index(currently_syncing) \
+        if currently_syncing else 0
+    stream_ids_to_sync = [cs.tap_stream_id for cs in ctx.catalog.streams
+                          if is_selected(cs)]
+    streams = [s for s in streams_.all_streams[start_idx:]
+               if s.tap_stream_id in stream_ids_to_sync]
+    for stream in streams:
+        ctx.state["currently_syncing"] = stream.tap_stream_id
+        output_schema(stream)
+        ctx.write_state()
+        stream.sync(ctx)
+    ctx.state["currently_syncing"] = None
+    ctx.write_state()
+
