+++ tap_zendesk_chat/utils.py
@@ -1,0 +1,21 @@
+#!/usr/bin/env python3
+import os
+import singer
+from singer.utils import load_json
+
+def get_abs_path(path):
+    return os.path.join(os.path.dirname(os.path.realpath(__file__)), path)
+
+
+def load_schema(tap_stream_id):
+    path = "schemas/{}.json".format(tap_stream_id)
+    schema = load_json(get_abs_path(path))
+    dependencies = schema.pop("tap_schema_dependencies", [])
+    refs = {}
+    for sub_stream_id in dependencies:
+        refs[sub_stream_id] = load_schema(sub_stream_id)
+    if refs:
+        singer.resolve_schema_references(schema, refs)
+    return schema
+
+
